Change the calibrate device screen to add the basic state management
The CalibrationWidget uses quiver.async CountDownTimer to count 10 seconds; during them the sensors data are collected and the biases are computed and stored
A state variable is used to keep track of what widget to show in the correct moment of the process.

ToDo: complete the sensor collection and bias computation code using streams, future and async; animate the svg picture of the phone.