Fix problem of calibration_helper.dart when app goes on pause
The calibration process is canceled every time the widget is disposed or the app goes on pause state, so we avoid listening to sensors when the flutter engine is not connected to any view.
Whenever we come back from another page or we resume the state is set to idle, unless previously it was done.