Remove PageView in main_screen.dart in favor of a simple list of screens to show
The page to show is choose from a list of screens every time the current index of the page changes; the list is initialized on init state so this ensure no duplicate screens are created.
The screen title on the app bar is a list of strings in string.dart

Note: No check is done on the index, but it's fine until the index is updated by the bottom navigation bar.