Change sensor_widget.dart to include the changes of sensor_monitor.dart
sensor_widget.dart contains a class SensorController that implements ChangeNotifier and has the purpose of convert what happens to sensor_monitor.dart into a state that SensorWidget will listen to and rebuild his children accordingly.
ToDo: include the functionality of cancel the controller if the app is on pause; this will require a flag for enabling and disabling this feature.