Change countdown_bloc to stop the measurement on close
With this commit the bloc countdown_bloc_impl.dart will automatically stop the pre measure countdown and the sensor monitor whenever he's closed.
Changed accordingly the measure_countdown's dispose and didPopRoute methods and the leave_confirmation_dialog, because now all they do is to call close() on the bloc; the leave dialog will return true or false and in didPopRoute the bloc is closed only if false.